requirements.txt
amqp==2.5.2
APScheduler==3.6.1
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.11
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.0.2
django-redis==4.11.0
django-reversion==3.0.7
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Login==0.4.1
future==0.18.2
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kombu==4.6.8
MarkupPy==1.14
MarkupSafe==1.1.1
mongoengine==0.18.2
odfpy==1.4.1
openpyxl==3.0.3
PyJWT==1.7.1
pymongo==3.9.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.1
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
